core: Update Annotations core: Handle updated Info message core: Handle new Settings message docs: Update Samples core: Update TouchPortalPluginAnnotationProcessor to add all fields matching entry.tp

+++ Annotations/src/main/java/com/github/ChristopheCVB/TouchPortal/Annotations/Plugin.java
@@ -49,9 +49,9 @@ public @interface Plugin {
     /**
      * Plugin version
      *
-     * @return String version
+     * @return long version
      */
-    int version();
+    long version();
 
     /**
      * Plugin colorDark
